refine babel/webpack integration
