+++ __tests__/__fixtures__/babel/webpack/actual.js
@@ -3,7 +3,7 @@ import imported from 'react-imported-component';
 
 const PreloadComponent = imported(() => import('./PreloadThis'));
 const PrefetchChunkComponent = imported(() => import('./ChunkThis'));
-const AsyncComponent0 = imported(() => import(/* webpackChunkName:namedChunk */'./MyComponent'));
+const AsyncComponent0 = imported(() => import(/* webpackChunkName: "namedChunk" */'./MyComponent'));
 
 const AsyncComponent1 = imported(() => import('./MyComponent'));
